Resolve GVN PRE issue with conversion of initialized loads to freeze poison
When PRE fails for a load that has been convert to freeze poison, the inserted
freeze poison must also be removed. If it is not then an infinite loop is
created due to the subsequent removal of the freeze poison in the next iteration
followed by the conversion of the load to freeze poison (which fails PRE
constraints again).