[llvm][Mem2Reg] Refactor util PromoteMemoryToRegister to use MemoryBuiltins
The promotion of memory to register function promoteSingleBlockAlloca is
refactored to call the MemoryBuiltins function getInitialValueOfAllocation. This
requires the passing of a TargetLibraryInfo object through function calls.

PromoteMemToReg uses a pointer to a constant TLI object. This done to make the
interface compatible with other optimizations that keep a const reference to
their TLI objects.

TODO:
* This commit only resolves compilation issues for the LLVMTransformUtils
  library.