[llvm][MemoryBuiltins] Add alloca support to getInitialValueOfAllocation
This commit is in support of future uninitialized memory handling and adds
alloca instruction support to getInitialValueOfAllocation. This unifies initial
memory state querying (both stack and heap) to a single utility function.

AMDGPU PromoteAlloca, Mem2Reg, GVN, and NewGVN optimizations are refactored to
take advantage of alloca support in getInitialValueOfAllocation.

Differential Revision: https://reviews.llvm.org/D155773