[SROA] Add freeze_bits support to inserted load instructions
Prevents new load instruction from using the current default enabling
freeze_bits by default.

Add MD_freeze_bits copy to:
- AllocaSliceRewriter.rewriteVectorizedLoadInst
- AllocaSliceRewriter.rewriteIntegerLoad

While this has no immediate effect functionally, The default will
eventually be to not apply freeze_bits by default. Thus the use of
copyMetadata will be required to preserve freeze_bits.

To resolve a segmentation fault encountered during the bootstrap build
process, rewriting integer loads always resulted in a new load
instruction with the !freeze_bits metadata applied
indiscriminately. The application of !freeze_bits is now more
judicious. The metadata is only applied if the source and target types
are equal. In such cases, the presence of !freeze_bits on the target
is determined by the state of the source load instruction. If the
types differ, !freeze_bits is applied to the target by default.