[SROA] Add !freeze support to inserted load instructions
Prevents new load instruction from using the current default enabling
freeze by default.

Add MD_freeze copy to:
- AllocaSliceRewriter.rewriteVectorizedLoadInst
- AllocaSliceRewriter.rewriteIntegerLoad

While this has no immediate effect functionally, The default will
eventually be to not apply freeze by default. Thus the use of
copyMetadata will be required to preserve !freeze.

To resolve a segmentation fault encountered during the bootstrap build process,
rewriting integer loads always resulted in a new load instruction with the
!freeze metadata applied indiscriminately. The application of !freeze is now
more judicious. The metadata is only applied if the source and target types are
equal. In such cases, the presence of !freeze on the target is determined by the
state of the source load instruction. If the types differ, !freeze is applied to
the target by default.