[SROA] Improve handling of !freeze_bits metadata in integer load rewriting
To resolve a segmentation fault encountered during the bootstrap build process,
rewriting integer loads always resulted in a new load instruction with the
!freeze_bits metadata applied indiscriminately. The application of !freeze_bits
is now more judicious. The metadata is only applied if the source and target
types are equal. In such cases, the presence of !freeze_bits on the target is
determined by the state of the source load instruction. If the types differ,
!freeze_bits is applied to the target by default.