[mem2reg][wip] Move freeze emission to store location
For allocs that have freeze_bits loads, emit freeze instruction at the site of
stores. This is part of the solution to the remove of arithmetic flags on loop
induction variables. The removal of arithmetic flags is caused by reduction of
freeze instructions in InstCombine, which prevents SLP from folding.

This commit also reduced freeze emission by using preexisting freeze when
transforming store instructions.